+++ debug.bash
@@ -1,0 +1,1 @@
+MallocGuardEdges=1 MallocScribble=1 MallocErrorAbort=1 RUST_LOG=rs_vst_host=debug cargo run --features debug-tools -- gui --malloc-debug